requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.0
asyncer==0.0.7
attrs==24.2.0
bidict==0.23.1
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
chainlit==1.2.0
charset-normalizer==3.3.2
chevron==0.14.0
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.7
Deprecated==1.2.14
distro==1.9.0
fastapi==0.112.4
filetype==1.2.0
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.66.1
grpcio-tools==1.66.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
importlib_metadata==8.4.0
Jinja2==3.1.4
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.5
langchain-openai==0.2.0
langchain-qdrant==0.1.4
langchain-text-splitters==0.3.0
langsmith==0.1.125
Lazify==0.4.0
literalai==0.0.607
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.8.2
nest-asyncio==1.6.0
numpy==1.26.4
openai==1.47.0
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
orjson==3.10.7
packaging==23.2
pandas==2.2.3
pillow==10.4.0
portalocker==2.10.1
protobuf==4.25.5
pyarrow==17.0.0
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.9.0
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-multipart==0.0.9
python-socketio==5.11.4
pytz==2024.2
pywin32==306
PyYAML==6.0.2
qdrant-client==1.11.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rich==13.8.1
rpds-py==0.20.0
simple-websocket==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.37.2
syncer==2.0.3
tenacity==8.5.0
tiktoken==0.7.0
toml==0.10.2
tomli==2.0.1
tornado==6.4.1
tqdm==4.66.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
uptrace==1.26.0
urllib3==2.2.3
uvicorn==0.25.0
watchdog==4.0.2
watchfiles==0.20.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.11.1
zipp==3.20.2
